requirements-vision.txt
# Vision MCP (install in a venv: pip install -r requirements-vision.txt)
mcp>=1.4.0
mss>=9.0.1
Pillow>=10.0.0
